@@ -3,24 +3,21 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "firebase": "^8.6.8",
-    "node-sass": "^6.0.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.4",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
+    "@reduxjs/toolkit": "^2.8.2",
+    "firebase": "^9.22.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.1.2",
+    "react-router-dom": "^6.23.0",
+    "react-scripts": "5.0.1",
     "react-stripe-checkout": "^2.6.3",
-    "redux": "^4.1.0",
+    "redux": "^4.2.1",
     "redux-logger": "^3.0.6",
     "redux-persist": "^6.0.0",
-    "redux-thunk": "^2.3.0",
-    "reselect": "^4.0.0",
-    "styled-components": "^5.3.0",
-    "web-vitals": "^1.0.1"
+    "redux-thunk": "^2.4.2",
+    "reselect": "^4.1.8",
+    "sass": "^1.70.0",
+    "styled-components": "^6.1.2"
   },
   "scripts": {
     "start": "react-scripts start",
