@@ -1,27 +1,21 @@
 import React from "react";
-
-import { ReactComponent as ShoppingIcon } from "../../assets/shopping-bag.svg";
+import { useDispatch, useSelector } from "react-redux";
 
 import "./cart-icon.styles.scss";
-import { createStructuredSelector } from "reselect";
 
-import { connect } from "react-redux";
 import { toggleCartHidden } from "../../redux/cart/cart.actions";
 import { selectCartItemsCount } from "../../redux/cart/cart.selectors";
 
-const CartIcon = ({ toggleCartHidden, itemCount }) => (
-  <div className='cart-icon' onClick={toggleCartHidden}>
-    <ShoppingIcon className='shopping-icon' />
-    <span className='item-count'>{itemCount}</span>
-  </div>
-);
+const CartIcon = () => {
+  const dispatch = useDispatch();
+  const itemCount = useSelector(selectCartItemsCount);
 
-const mapDispatchToProps = (dispatch) => ({
-  toggleCartHidden: () => dispatch(toggleCartHidden()),
-});
+  return (
+    <div className="cart-icon" onClick={() => dispatch(toggleCartHidden())}>
+      <img src="/assets/shopping-bag.svg" alt="Shopping Icon" className="shopping-icon" />
+      <span className="item-count">{itemCount}</span>
+    </div>
+  );
+};
 
-const mapStateToProps = createStructuredSelector({
-  itemCount: selectCartItemsCount,
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(CartIcon);
+export default CartIcon;
