@@ -1,22 +1,17 @@
 import React from "react";
-import { connect } from "react-redux";
-import { createStructuredSelector } from "reselect";
-
+import { useSelector } from "react-redux";
 import { selectDirectorySections } from "../../redux/directory/directory.selectors";
-
 import "./directory.styles.scss";
 import MenuItem from "../menu-item/menu-item.component";
 
-const Directory = ({ sections }) => (
+const Directory = () => {
+  const sections = useSelector(selectDirectorySections);
+  return (
   <div className='directory-menu'>
     {sections.map(({ id, ...otherSectionProps }) => (
       <MenuItem key={id} {...otherSectionProps} />
     ))}
   </div>
-);
+);}
 
-const mapStateToProps = createStructuredSelector({
-  sections: selectDirectorySections,
-});
-
-export default connect(mapStateToProps)(Directory);
+export default Directory;
