@@ -1,6 +1,5 @@
 import React from "react";
-import { connect } from "react-redux";
-import { createStructuredSelector } from "reselect";
+import { useSelector } from "react-redux";
 
 import CheckoutItem from "../../components/checkout-item/checkout-item.component";
 import StripeCheckoutButton from "../../components/stripe-button/stripe-button.component";
@@ -12,45 +11,48 @@ import {
 
 import "./checkout.styles.scss";
 
-const CheckoutPage = ({ cartItems, total }) => (
-  <div className='checkout-page'>
-    <div className='checkout-header'>
-      <div className='header'>
-        <span>Product</span>
+const CheckoutPage = () => {
+  const cartItems = useSelector(selectCartItems);
+  const total = useSelector(selectCartTotal);
+
+  return (
+    <div className='checkout-page'>
+      <div className='checkout-header'>
+        <div className='header'>
+          <span>Product</span>
+        </div>
+        <div className='header'>
+          <span>Description</span>
+        </div>
+        <div className='header'>
+          <span>Quantity</span>
+        </div>
+        <div className='header'>
+          <span>Price</span>
+        </div>
+        <div className='header'>
+          <span>Remove</span>
+        </div>
       </div>
-      <div className='header'>
-        <span>Description</span>
+      {cartItems.map((cartItem) => (
+        <CheckoutItem key={cartItem.id} cartItem={cartItem} />
+      ))}
+      <div className='total'>TOTAL: ${total}</div>
+
+      <div className='test-warning'>
+        *Please use the following test credit card for payments*
+        <br />
+        4242 4242 4242 4242 - Exp: any future date - CVV: any 3 digits
+        <br />
+        Refer this Link:{" "}
+        <a href="https://stripe.com/docs/testing" target="_blank" rel="noreferrer">
+          https://stripe.com/docs/testing
+        </a>
       </div>
-      <div className='header'>
-        <span>Quantity</span>
-      </div>
-      <div className='header'>
-        <span>Price</span>
-      </div>
-      <div className='header'>
-        <span>Remove</span>
-      </div>
+
+      <StripeCheckoutButton price={total} />
     </div>
-    {cartItems.map((cartItem) => (
-      <CheckoutItem key={cartItem.id} cartItem={cartItem} />
-    ))}
-    <div className='total'>TOTAL: ${total}</div>
+  );
+};
 
-    <div className='test-warning'>
-      *Please use the following test credi card for payments*
-      <br />
-      4242 4242 4242 4242 - Exp: anyfuture date - CVV: any 3 digits
-      <br />
-      Refer this Link: https://stripe.com/docs/testing
-    </div>
-
-    <StripeCheckoutButton price={total} />
-  </div>
-);
-
-const mapStateToProps = createStructuredSelector({
-  cartItems: selectCartItems,
-  total: selectCartTotal,
-});
-
-export default connect(mapStateToProps)(CheckoutPage);
+export default CheckoutPage;
