@@ -1,18 +1,24 @@
 import React from "react";
-import { connect } from "react-redux";
+import { useParams } from "react-router-dom";
+import { useSelector } from "react-redux";
 
 import CollectionItem from "../../components/collection-item/collection-item.component";
-
 import { selectCollection } from "../../redux/shop/shop.selectors";
 
 import "./collection.styles.scss";
 
-const CollectionPage = ({ collection }) => {
+const CollectionPage = () => {
+  const { collectionId } = useParams();
+  const collection = useSelector(selectCollection(collectionId));
+
+  if (!collection) return <div>Collection not found.</div>;
+
   const { title, items } = collection;
+
   return (
-    <div className='collection-page'>
-      <h2 className='title'>{title}</h2>
-      <div className='items'>
+    <div className="collection-page">
+      <h2 className="title">{title}</h2>
+      <div className="items">
         {items.map((item) => (
           <CollectionItem key={item.id} item={item} />
         ))}
@@ -21,8 +27,4 @@ const CollectionPage = ({ collection }) => {
   );
 };
 
-const mapStateToProps = (state, ownProps) => ({
-  collection: selectCollection(ownProps.match.params.collectionId)(state),
-});
-
-export default connect(mapStateToProps)(CollectionPage);
+export default CollectionPage;
