@@ -2,10 +2,11 @@ import React from "react";
 import Directory from "../../components/directory/directory.component";
 import { HomePageContainer } from "./homepage.styles";
 
-const Homepage = () => (
+const Homepage = () => {
+  return (
   <HomePageContainer>
     <Directory />
   </HomePageContainer>
-);
+)};
 
 export default Homepage;
