@@ -1,7 +1,6 @@
-import React from "react";
-import { Route } from "react-router-dom";
-import { createStructuredSelector } from "reselect";
-import { connect } from "react-redux";
+import React, { useEffect } from "react";
+import { Route, useLocation, useMatch, useParams } from "react-router-dom";
+import { useDispatch, useSelector } from "react-redux";
 
 import { fetchCollectionsStartAsync } from "../../redux/shop/shop.actions";
 import {
@@ -10,53 +9,33 @@ import {
 } from "../../redux/shop/shop.selectors";
 
 import WithSpinner from "../../components/with-spinner/with-spinner.component";
-
 import CollectionsOverview from "../../components/collections-overview/collections-overview.component";
 import CollectionPage from "../collection/collection.component";
 
 const CollectionsOverviewWithSpinner = WithSpinner(CollectionsOverview);
 const CollectionPageWithSpinner = WithSpinner(CollectionPage);
 
-class ShopPage extends React.Component {
-  componentDidMount() {
-    const { fetchCollectionsStartAsync } = this.props;
-    fetchCollectionsStartAsync();
-  }
-  render() {
-    const { match, isCollectionFetching, isCollectionsLoaded } = this.props;
-    return (
-      <div className='shop-page'>
-        <Route
-          exact
-          path={`${match.path}`}
-          render={(props) => (
-            <CollectionsOverviewWithSpinner
-              isLoading={isCollectionFetching}
-              {...props}
-            />
-          )}
-        />
-        <Route
-          path={`${match.path}/:collectionId`}
-          render={(props) => (
-            <CollectionPageWithSpinner
-              isLoading={!isCollectionsLoaded}
-              {...props}
-            />
-          )}
-        />
-      </div>
-    );
-  }
-}
+const ShopPage = () => {
+  const { collectionId } = useParams();
+  const location = useLocation();
+  const dispatch = useDispatch();
 
-const mapStateToProps = createStructuredSelector({
-  isCollectionFetching: selectIsCollectionFetching,
-  isCollectionsLoaded: selectIsCollectionsLoaded,
-});
+  const isCollectionFetching = useSelector(selectIsCollectionFetching);
+  const isCollectionsLoaded = useSelector(selectIsCollectionsLoaded);
 
-const mapDispatchToProps = (dispatch) => ({
-  fetchCollectionsStartAsync: () => dispatch(fetchCollectionsStartAsync()),
-});
+  useEffect(() => {
+    dispatch(fetchCollectionsStartAsync());
+  }, [dispatch]);
 
-export default connect(mapStateToProps, mapDispatchToProps)(ShopPage);
+  return (
+    <div className="shop-page">
+      {!collectionId ? (
+        <CollectionsOverviewWithSpinner isLoading={isCollectionFetching} />
+      ) : (
+        <CollectionPageWithSpinner isLoading={!isCollectionsLoaded} />
+      )}
+    </div>
+  );
+};
+
+export default ShopPage;
