@@ -1,23 +1,25 @@
-import { combineReducers } from "redux";
-import { persistReducer } from "redux-persist";
-import storage from "redux-persist/lib/storage";
+import { combineReducers } from 'redux';
+import { persistReducer } from 'redux-persist';
+import storage from 'redux-persist/lib/storage';
 
-import CartReducer from "./cart/cart.reducer";
-import userReducer from "./user/user.reducer";
-import directoryReducer from "./directory/directory.reducer";
-import shopReducer from "./shop/shop.reducer";
+import cartReducer from './cart/cart.reducer';
+import userReducer from './user/user.reducer';
+import directoryReducer from './directory/directory.reducer';
+import shopReducer from './shop/shop.reducer';
 
 const persistConfig = {
-  key: "root",
+  key: 'root',
   storage,
-  whitelist: ["cart"],
+  whitelist: ['cart'], // you can add 'user' if needed
 };
 
 const rootReducer = combineReducers({
   user: userReducer,
-  cart: CartReducer,
+  cart: cartReducer,
   directory: directoryReducer,
   shop: shopReducer,
 });
 
-export default persistReducer(persistConfig, rootReducer);
+const persistedReducer = persistReducer(persistConfig, rootReducer);
+
+export default persistedReducer;
