@@ -1,17 +1,18 @@
+import { collection, getDocs } from "firebase/firestore";
 import ShopActionTypes from "./shop.types";
-
 import {
   firestore,
   convertCollectionsSnapshotToMap,
 } from "../../firebase/firebase.utils";
 
+// Action creators
 export const fetchCollectionsStart = () => ({
   type: ShopActionTypes.FETCH_COLLECTIONS_START,
 });
 
-export const fetchCollectionsStuccess = (collectionMap) => ({
+export const fetchCollectionsSuccess = (collectionsMap) => ({
   type: ShopActionTypes.FETCH_COLLECTIONS_SUCCESS,
-  payload: collectionMap,
+  payload: collectionsMap,
 });
 
 export const fetchCollectionsFailure = (errorMessage) => ({
@@ -19,18 +20,19 @@ export const fetchCollectionsFailure = (errorMessage) => ({
   payload: errorMessage,
 });
 
+// Thunk async action
 export const fetchCollectionsStartAsync = () => {
-  return (dispatch) => {
-    const collectionsRef = firestore.collection("collections");
+  return async (dispatch) => {
     dispatch(fetchCollectionsStart());
 
-    collectionsRef
-      .get()
-      .then((snapshot) => {
-        const collectionsMap = convertCollectionsSnapshotToMap(snapshot);
-        // updateCollections(collectionsMap);
-        dispatch(fetchCollectionsStuccess(collectionsMap));
-      })
-      .catch((error) => dispatch(fetchCollectionsFailure(error.message)));
+    try {
+      const collectionRef = collection(firestore, "collections");
+      const snapshot = await getDocs(collectionRef);
+      const collectionsMap = convertCollectionsSnapshotToMap(snapshot);
+      dispatch(fetchCollectionsSuccess(collectionsMap));
+    } catch (error) {
+      console.log("error: ", error);
+      dispatch(fetchCollectionsFailure(error.message));
+    }
   };
 };
